plugins: extract schemas to schemas.py
Why:

* comply with wazo-notebook/file-structure.md
