+++ wazo_calld/plugins/voicemails/http.py
@@ -5,9 +5,6 @@ import re
 
 from flask import request
 from flask import Response
-from marshmallow import Schema, fields
-
-from xivo.mallow.validate import OneOf
 
 from wazo_calld.auth import get_token_user_uuid_from_request
 from wazo_calld.auth import required_acl
@@ -19,52 +16,14 @@ from .exceptions import (
     NoSuchVoicemailGreeting,
     InvalidVoicemailMessageID,
 )
-
-VALID_GREETINGS = ["unavailable", "busy", "name"]
-
-
-class VoicemailMessageBaseSchema(Schema):
-    id = fields.String()
-    caller_id_name = fields.String()
-    caller_id_num = fields.String()
-    duration = fields.Integer()
-    timestamp = fields.Integer()
-
-
-class VoicemailFolderBaseSchema(Schema):
-    id = fields.Integer()
-    name = fields.String()
-    type = fields.String()
-
-
-class VoicemailMessageSchema(VoicemailMessageBaseSchema):
-    folder = fields.Nested(VoicemailFolderBaseSchema)
-
-
-class VoicemailFolderSchema(VoicemailFolderBaseSchema):
-    messages = fields.Nested(VoicemailMessageBaseSchema, many=True)
-
-
-class VoicemailSchema(Schema):
-    id = fields.Integer()
-    name = fields.String()
-    number = fields.String()
-    folders = fields.Nested(VoicemailFolderSchema, many=True)
-
-
-class VoicemailMessageUpdateSchema(Schema):
-    folder_id = fields.Integer(required=True)
-
-
-class VoicemailGreetingCopySchema(Schema):
-    dest_greeting = fields.String(validate=OneOf(VALID_GREETINGS))
-
-
-voicemail_schema = VoicemailSchema()
-voicemail_folder_schema = VoicemailFolderSchema()
-voicemail_message_schema = VoicemailMessageSchema()
-voicemail_message_update_schema = VoicemailMessageUpdateSchema()
-voicemail_greeting_copy_schema = VoicemailGreetingCopySchema()
+from .schemas import (
+    VALID_GREETINGS,
+    voicemail_schema,
+    voicemail_folder_schema,
+    voicemail_message_schema,
+    voicemail_message_update_schema,
+    voicemail_greeting_copy_schema,
+)
 
 
 class _BaseVoicemailResource(AuthResource):
